Make test cases consistent and correct.
cURL commands in the wild are typically meant to be pasted into the
console, and thus have quoting for a shell.  This is done inconsistently
in the test cases, but ought to be done all the time when it makes a
difference (e.g., when ? or & appear in the URL).

It also came to light that the get_with_form test case is syntactically
incorrect, and the corresponding test file did not have a good expected
value.